pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hftool"
version = "0.2.0"
description = "A CLI for Hugging Face models: text-to-image, text-to-video, TTS, ASR, and more."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT License"}
authors = [
  {name = "zashboy", email = "zashboy@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Environment :: Console",
    "Typing :: Typed",
]

dependencies = [
    "transformers>=4.45.0",
    "torch>=2.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
# Image handling (for image inputs/outputs)
with_image = ["Pillow>=9.0.0"]

# Audio handling (for TTS/ASR)
with_audio = ["soundfile>=0.12.0"]

# HTTP support (for URL inputs)
with_http = ["requests>=2.20.0"]

# Text-to-Image generation (Z-Image, SDXL, FLUX)
with_t2i = [
    "diffusers>=0.32.0",
    "Pillow>=9.0.0",
    "accelerate>=0.26.0",
]

# Text-to-Video generation (HunyuanVideo, CogVideoX, Wan2.2)
# Requires system ffmpeg for video encoding
with_t2v = [
    "diffusers>=0.32.0",
    "Pillow>=9.0.0",
    "accelerate>=0.26.0",
]

# Text-to-Speech (VibeVoice, Bark, MMS-TTS)
# Requires system ffmpeg for MP3 conversion
with_tts = [
    "transformers>=4.45.0",
    "soundfile>=0.12.0",
]

# Speech-to-Text / ASR (Whisper)
with_stt = [
    "transformers>=4.45.0",
    "soundfile>=0.12.0",
]

# Advanced audio processing (resampling, feature extraction)
with_audio_advanced = [
    "soundfile>=0.12.0",
    "librosa>=0.9.0",
    "scipy",
]

# Advanced vision processing
with_vision_advanced = [
    "Pillow>=9.0.0",
    "opencv-python-headless>=4.5.0",
]

# Model quantization (requires CUDA/ROCm)
with_quantization = ["bitsandbytes>=0.39.0"]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

# All dependencies (for full functionality)
all = [
    # Image
    "Pillow>=9.0.0",
    # Audio
    "soundfile>=0.12.0",
    # HTTP
    "requests>=2.20.0",
    # Diffusers (T2I, T2V)
    "diffusers>=0.32.0",
    "accelerate>=0.26.0",
    # Audio advanced
    "librosa>=0.9.0",
    "scipy",
    # Vision advanced
    "opencv-python-headless>=4.5.0",
]

[project.scripts]
hftool = "hftool.cli:main"

[project.urls]
Homepage = "https://github.com/zashboy-websites/hftool"
Repository = "https://github.com/zashboy-websites/hftool"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["."]
include = ["hftool*"]
